refactor(all): 디렉터리 구조 변경
전체 디렉토리 구조 변경 및 테스트 가독성 향상
