+++ src/main/java/lotto/domain/model/lotto/generator/LottoAutoGenerator.java
@@ -1,6 +1,7 @@
-package lotto.domain.model.lotto;
+package lotto.domain.model.lotto.generator;
 
 import camp.nextstep.edu.missionutils.Randoms;
+import lotto.domain.model.user.Lotto;
 
 import java.util.List;
 import java.util.stream.IntStream;
