+++ src/main/java/lotto/domain/model/lotto/generator/LottoGenerator.java
@@ -1,4 +1,6 @@
-package lotto.domain.model.lotto;
+package lotto.domain.model.lotto.generator;
+
+import lotto.domain.model.user.Lotto;
 
 import java.util.List;
