+++ src/test/java/lotto/LottoTest.java
@@ -1,10 +1,11 @@
 package lotto;
 
 import lotto.common.constant.PrintFormatConst;
-import lotto.domain.model.lotto.Lotto;
+import lotto.domain.model.user.Lotto;
 import lotto.domain.utils.TestLotto;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
@@ -26,31 +27,40 @@ class LottoTest {
                 .isInstanceOf(IllegalArgumentException.class);
     }
 
-    @DisplayName("올바른 로또 번호를 형식에 맞게 반환한다.")
-    @Test
-    void printTest() {
-        //given
-        String format1 = "1, 2, 3, 4, 5, 6";
-        String format2 = "6, 5, 4, 3, 2, 1";
 
-        Lotto lotto1 = TestLotto.createTestLotto(List.of(1, 2, 3, 4, 5, 6));
-        Lotto lotto2 = TestLotto.createTestLotto(List.of(6, 5, 4, 3, 2, 1));
+    @Nested
+    @DisplayName("print 메서드는")
+    class PrintTest {
+        @DisplayName("올바른 로또 번호를 형식에 맞게 반환한다.")
+        @Test
+        void printTest() {
+            //given
+            String format1 = "1, 2, 3, 4, 5, 6";
+            String format2 = "6, 5, 4, 3, 2, 1";
+
+            Lotto lotto1 = TestLotto.createTestLotto(List.of(1, 2, 3, 4, 5, 6));
+            Lotto lotto2 = TestLotto.createTestLotto(List.of(6, 5, 4, 3, 2, 1));
 
 
-        //when
-        String print1 = lotto1.print();
-        String print2 = lotto2.print();
+            //when
+            String print1 = lotto1.print();
+            String print2 = lotto2.print();
 
-        //then
-        Assertions.assertThat(print1).isEqualTo(
-                String.format(PrintFormatConst.LOTTO_NUMBERS_FORMAT,
-                        format1
-                )
-        );
-        Assertions.assertThat(print2).isEqualTo(
-                String.format(PrintFormatConst.LOTTO_NUMBERS_FORMAT,
-                        format2
-                )
-        );
+            //then
+            Assertions.assertThat(print1).isEqualTo(
+                    String.format(PrintFormatConst.LOTTO_NUMBERS_FORMAT,
+                            format1
+                    )
+            );
+            Assertions.assertThat(print2).isEqualTo(
+                    String.format(PrintFormatConst.LOTTO_NUMBERS_FORMAT,
+                            format2
+                    )
+            );
+        }
     }
+
+
+
+
 }
