+++ src/test/java/lotto/domain/model/user/LottoRankTest.java
@@ -1,10 +1,8 @@
-package lotto.domain.model.lotto;
+package lotto.domain.model.user;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 class LottoRankTest {
